% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{firstButton}
\alias{firstButton}
\title{Code for the default "first" control}
\usage{
firstButton(label, class = c("btn", "btn-default"))
}
\arguments{
\item{label}{control label.}

\item{class}{control CSS classes. \code{first-screen} is automatically added.}
}
\description{
This generates the code of the default "first" control, which is only displayed
on the first screen of a glkide, and can be used in custom controls.
}
\seealso{
nextButton prevButton lastButton
}
