% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{glideControls}
\alias{glideControls}
\title{Default controls layout}
\usage{
glideControls(previous_content, next_content, jump_content = NULL)
}
\arguments{
\item{previous_content}{Content of the "previous" (left) zone.}

\item{next_content}{Content of the "next" (right) zone.}
}
\description{
Creates an horizontal layout with both "previous" and "next" contents
side by side.
}
\examples{
glideControls(
  prevButton("Back"),
  list(
    lastButton(href = "https://example.com", "Go to website"),
    nextButton("Next")
  )
)
}
