% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{lastButton}
\alias{lastButton}
\title{Code for the default "last" control}
\usage{
lastButton(label, class = c("btn", "btn-success"))
}
\arguments{
\item{label}{control label.}

\item{class}{control CSS classes. \code{last-screen} is automatically added.}
}
\description{
This generates the code of the default "last" control, which is only displayed
on the last screen of a glide, and can be used in custom controls.
}
\seealso{
nextButton prevButton firstButton
}
