% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{prevButton}
\alias{prevButton}
\title{Code for the default "back" control}
\usage{
prevButton(label = paste(shiny::icon("chevron-left", lib = "glyphicon"),
  "Back"), class = c("btn", "btn-default"))
}
\arguments{
\item{label}{control label.}

\item{class}{control CSS classes. \code{prev-screen} is automatically added.}
}
\description{
This generates the code of the default "back" control, and can be used in custom
controls.
}
\seealso{
nextButton firstButton lastButton
}
