% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{firstButton}
\alias{firstButton}
\alias{lastButton}
\title{Create a glide control only shown on first or last screen}
\usage{
firstButton(class = c("btn", "btn-default"), ...)

lastButton(class = c("btn", "btn-success"), ...)
}
\arguments{
\item{class}{CSS classes of the control. The needed class is automatically added.}

\item{...}{content of the control}
}
\description{
Create a glide control only shown on first or last screen
}
