% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{firstButton}
\alias{firstButton}
\alias{firstBtn}
\alias{lastButton}
\alias{lastBtn}
\title{Create a glide control only shown on first or last screen}
\usage{
firstButton(class = c("btn", "btn-default"), ...)

firstBtn(inputId, label, icon = NULL, width = NULL, ...)

lastButton(class = c("btn", "btn-success"), ...)

lastBtn(
  inputId,
  label,
  icon = icon("ok", lib = "glyphicon"),
  width = NULL,
  ...
)
}
\arguments{
\item{class}{CSS classes of the control. The needed class is automatically added.}

\item{...}{content of the control}
}
\description{
Create a glide control only shown on first or last screen
}
\details{
These controls generate an \verb{<a>} tag, so you can use \code{href} attributes.

\verb{firstButton``is only shown on the first screen of the app, and }finalButton` only
on the last screen.
}
\examples{
firstButton("Go to website", href = "https://example.com", class = "btn btn-primary")

}
