% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{nextButton}
\alias{nextButton}
\alias{prevButton}
\title{Code for the default controls}
\usage{
nextButton(label = paste("Next", shiny::icon("chevron-right", lib =
  "glyphicon")), class = c("btn", "btn-primary"))

prevButton(label = paste(shiny::icon("chevron-left", lib = "glyphicon"),
  "Back"), class = c("btn", "btn-default"))
}
\arguments{
\item{label}{control label.}

\item{class}{control CSS classes. The needed class is automatically added.}
}
\description{
This generates the code of the default controls, and can be used in custom
controls.
}
\details{
\code{prevButton} is hidden on the first screen, while \code{nextButton} is hidden on
the last one.
}
