% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{nextButton}
\alias{nextButton}
\alias{prevButton}
\title{Code for the default controls}
\usage{
nextButton(class = c("btn", "btn-primary"))

prevButton(class = c("btn", "btn-default"))
}
\arguments{
\item{class}{control CSS classes. The needed class is automatically added.}
}
\description{
This generates the code of the default controls, and can be used in custom
controls.
}
\details{
\code{prevButton} is hidden on the first screen, while \code{nextButton} is hidden on
the last one.
The buttons labels are set with the \code{next_label} and \code{previous_label}
arguments of \code{glide()}.
}
\seealso{
glide
}
