% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{nextButton}
\alias{nextButton}
\alias{nextBtn}
\alias{prevButton}
\alias{prevBtn}
\title{Code for the default controls}
\usage{
nextButton(class = c("btn", "btn-primary"))

nextBtn(inputId, label, icon = NULL, width = NULL, ...)

prevButton(class = c("btn", "btn-default"))

prevBtn(inputId, label, icon = NULL, width = NULL, ...)
}
\arguments{
\item{class}{control CSS classes. The needed class is automatically added.}

\item{inputId, label, icon, width, ...}{see decription in \code{\link[shiny]{actionButton}}}
}
\description{
This generates the code of the default controls, and can be used in custom
controls.
}
\details{
\code{prevButton} is hidden on the first screen, while \code{nextButton} is hidden on
the last one.
The buttons labels are set with the \code{next_label} and \code{previous_label}
arguments of \code{glide()}.
}
\seealso{
glide
}
